Split AccessibleObject
